# gateway/docs/swagger.yaml
basePath: /api
definitions:
  models.AuthResponse:
    properties:
      token:
        type: string
      user:
        $ref: '#/definitions/models.UserPublic'
    type: object
  models.CreateOrderItem:
    properties:
      product_id:
        type: string
      quantity:
        minimum: 1
        type: integer
    required:
    - product_id
    - quantity
    type: object
  models.CreateOrderRequest:
    properties:
      items:
        items:
          $ref: '#/definitions/models.CreateOrderItem'
        minItems: 1
        type: array
      notes:
        type: string
      shipping_address:
        $ref: '#/definitions/models.ShippingAddress'
    required:
    - items
    - shipping_address
    type: object
  models.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  models.Order:
    properties:
      created_at:
        type: string
      id:
        type: string
      items:
        items:
          $ref: '#/definitions/models.OrderItem'
        type: array
      notes:
        type: string
      shipping_address:
        $ref: '#/definitions/models.ShippingAddress'
      shipping_cost:
        type: number
      status:
        $ref: '#/definitions/models.OrderStatus'
      subtotal:
        type: number
      total:
        type: number
      updated_at:
        type: string
      user_id:
        type: string
    type: object
  models.OrderItem:
    properties:
      line_total:
        type: number
      product_id:
        type: string
      product_name:
        type: string
      product_type:
        type: string
      quantity:
        type: integer
      unit_price:
        type: number
    type: object
  models.OrderListResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/models.Order'
        type: array
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
    type: object
  models.OrderStatus:
    enum:
    - pending
    - payment_processing
    - payment_failed
    - confirmed
    - preparing
    - shipped
    - in_transit
    - delivered
    - cancelled
    - refund_pending
    - refunded
    type: string
    x-enum-varnames:
    - StatusPending
    - StatusPaymentProcessing
    - StatusPaymentFailed
    - StatusConfirmed
    - StatusPreparing
    - StatusShipped
    - StatusInTransit
    - StatusDelivered
    - StatusCancelled
    - StatusRefundPending
    - StatusRefunded
  models.Product:
    properties:
      attributes:
        type: object
      created_at:
        type: string
      group:
        $ref: '#/definitions/models.ProductGroup'
      id:
        type: string
      image_url:
        type: string
      in_stock:
        type: boolean
      name:
        type: string
      price:
        type: number
      stock_count:
        type: integer
      type:
        $ref: '#/definitions/models.ProductType'
      updated_at:
        type: string
    type: object
  models.ProductGroup:
    enum:
    - structural
    - guidance
    - payload
    - propulsion
    type: string
    x-enum-varnames:
    - GroupStructural
    - GroupGuidance
    - GroupPayload
    - GroupPropulsion
  models.ProductListItem:
    properties:
      group:
        $ref: '#/definitions/models.ProductGroup'
      id:
        type: string
      image_url:
        type: string
      in_stock:
        type: boolean
      name:
        type: string
      price:
        type: number
      stock_count:
        type: integer
      type:
        $ref: '#/definitions/models.ProductType'
    type: object
  models.ProductListResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/models.ProductListItem'
        type: array
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
    type: object
  models.ProductType:
    enum:
    - rocket_frame
    - panels_fuselage
    - control_fins
    - flight_computer
    - nav_sensors
    - control_actuation
    - telemetry
    - nose_cone
    - crewed_cabin
    - cargo_module
    - liquid_engine
    - propellant_tank
    - rocket_nozzle
    type: string
    x-enum-varnames:
    - TypeRocketFrame
    - TypePanelsFuselage
    - TypeControlFins
    - TypeFlightComputer
    - TypeNavSensors
    - TypeControlActuation
    - TypeTelemetry
    - TypeNoseCone
    - TypeCrewedCabin
    - TypeCargoModule
    - TypeLiquidEngine
    - TypePropellantTank
    - TypeRocketNozzle
  models.RegisterRequest:
    properties:
      email:
        type: string
      name:
        minLength: 2
        type: string
      password:
        minLength: 8
        type: string
    required:
    - email
    - name
    - password
    type: object
  models.ShippingAddress:
    properties:
      address_line_1:
        type: string
      address_line_2:
        type: string
      city:
        type: string
      country:
        type: string
      facility_name:
        type: string
      postal_code:
        type: string
      site_code:
        type: string
    required:
    - address_line_1
    - city
    - country
    - facility_name
    - postal_code
    - site_code
    type: object
  models.UserPublic:
    properties:
      created_at:
        type: string
      email:
        type: string
      id:
        type: string
      name:
        type: string
    type: object
host: localhost:8000
info:
  contact: {}
  description: API gateway for the Starbound rocket parts e-commerce platform
  title: Starbound API
  version: "1.0"
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: Authenticates a user and returns a JWT token
      parameters:
      - description: Login credentials
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.AuthResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Login
      tags:
      - auth
  /auth/logout:
    post:
      description: Stateless logout — client should discard the token
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Logout
      tags:
      - auth
  /auth/me:
    get:
      description: Returns the authenticated user's profile from the JWT
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.UserPublic'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get current user
      tags:
      - auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Creates a new user account and returns a JWT token
      parameters:
      - description: Registration details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.AuthResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Register a new account
      tags:
      - auth
  /orders:
    get:
      description: Returns a paginated list of all orders for the authenticated user
      parameters:
      - description: Page number (default 1)
        in: query
        name: page
        type: integer
      - description: Items per page (default 20)
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.OrderListResponse'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List current user's orders
      tags:
      - orders
    post:
      consumes:
      - application/json
      description: Creates a new order for the authenticated user
      parameters:
      - description: Order details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.CreateOrderRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Order'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "422":
          description: Unprocessable Entity
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create a new order
      tags:
      - orders
  /orders/{id}:
    get:
      description: Returns a single order by ID — user can only access their own orders
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Order'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get a single order
      tags:
      - orders
  /orders/{id}/cancel:
    put:
      description: Cancels a pending order — only works for cancellable statuses
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Order'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Cancel an order
      tags:
      - orders
  /products:
    get:
      description: Returns a paginated list of products, optionally filtered by group,
        type, or search term
      parameters:
      - description: Product group
        enum:
        - structural
        - guidance
        - payload
        - propulsion
        in: query
        name: group
        type: string
      - description: Product type
        enum:
        - rocket_frame
        - panels_fuselage
        - control_fins
        - flight_computer
        - nav_sensors
        - control_actuation
        - telemetry
        - nose_cone
        - crewed_cabin
        - cargo_module
        - liquid_engine
        - propellant_tank
        - rocket_nozzle
        in: query
        name: type
        type: string
      - description: Search by name
        in: query
        name: search
        type: string
      - description: Page number (default 1)
        in: query
        name: page
        type: integer
      - description: Items per page (default 20, max 100)
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ProductListResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List products
      tags:
      - products
  /products/{id}:
    get:
      description: Returns a single product with full attributes by ID
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Product'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get a single product
      tags:
      - products
  /products/groups:
    get:
      description: Returns the full list of product groups and their types
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Get product taxonomy
      tags:
      - products
securityDefinitions:
  BearerAuth:
    description: 'JWT token — format: Bearer <token>'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"